// stay-in-sync-configurator-ui/src/app/features/source-system/models/sourceSystem.ts
/**
 * core-management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SourceSystemApiRequestConfiguration } from './sourceSystemApiRequestConfiguration';
import { SourceSystemEndpoint } from './sourceSystemEndpoint';
import { ApiHeader } from './apiHeader';
import { SyncSystemAuthConfig } from './syncSystemAuthConfig';
import { SyncSystemEndpoint } from './syncSystemEndpoint';


export interface SourceSystem { 
    id?: number;
    name?: string;
    apiUrl?: string;
    description?: string;
    apiType?: string;
    openApiSpec?: Blob;
    syncSystemEndpoints?: Set<SyncSystemEndpoint>;
    apiRequestHeaders?: Set<ApiHeader>;
    authConfig?: SyncSystemAuthConfig;
    sourceSystemEndpoints?: Set<SourceSystemEndpoint>;
    sourceSystemApiRequestConfigurations?: Set<SourceSystemApiRequestConfiguration>;
}